vkd3d: Ignore application use of CREATE_NOT_ZEROED.
It seems like Windows drivers ignore this flag and as a result,
games seem to be buggy in this area. The only rational solution is to
always clear. For non-suballocated resources, we should always be able
to rely on the kernel doing this for us, with the unfortunate
exception of currently bugged amdgpu kernels ...

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>