vkd3d: Deal with instantaneous signal + rewinds.
It's possible that signals may get lost.
Implement a ticketing system which is registered at
ID3D12CommandQueue::Wait() time on caller thread.

Tests demonstrate that the instant any queued WAIT is satisfied,
it is unblocked forever, even if there are other dependencies which
block the execution of said WAIT.

This resolves some potential deadlocks where there are pulsing signals
and rewinds.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>